pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cb2game"
version = "0.9.1"
authors = [
  { name="Jacob Sharf", email="jacobsharf@gmail.com" },
  { name="Mustafa Omer Gul", email="momergul@gmail.com" },
  { name="Yoav Artzi", email="yoav@cs.cornell.edu" },
]
description = "A Collaborative Natural Language Instruction Game"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "aiohttp==3.8.4",
  "fire==0.4.0",
  "numpy==1.23.5",
  "python-dateutil==2.8.2",
  "dataclasses_json==0.5.6",
  "pygame==2.1.3.dev8",
  "peewee==3.14.8",
  "freetype-py==2.2.0",
  "matplotlib==3.5.1",
  "humanhash3==0.0.6",
  "orjson==3.6.7",
  "pendulum==2.1.2",
  "mashumaro==3.5",
  "viztracer==0.15.6",
  "tqdm==4.64.0",
  "sparklines==0.4.2",
  "gymnasium==0.26.3",
  "requests==2.28.1",
  "pyyaml==6.0",
  "colorhash==1.1.0",
  "pre-commit==2.20.0",
  "autoflake==1.7.7",
  "pylint==2.15.5",
  "google-auth==2.14.0",
  "nest_asyncio==1.5.6",
  "time-machine==2.8.2",
  "appdirs==1.4.4",
  "cachecontrol==0.12.11",
  "openai==0.27.2",
  "evals==0.1.1",
  "pdoc==13.1.1",
  "gitpython==3.1.31",
  "aiohttp_session==2.12.0",
  "aiohttp_session[secure]==2.12.0",
]

[project.urls]
"Homepage" = "https://cb2.ai/"
"Bug Tracker" = "https://github.com/lil-lab/cb2/issues"

[tool.isort]
profile = "black"

[tool.autoflake]
remove-all-unused-imports=true
remove-duplicate-keys=true

[tool.black]
safe=true # Makes sure AST is semantically equivalent.
line-length=88
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/game/*.py    # Exclude Unity directory.
  | ^/env/*.py    # Exclude gym env for now.
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''
